compound CVX rewards 🚀

+++ src/interfaces/ICurvePool.sol
@@ -14,7 +14,7 @@ interface ICurvePool {
         payable
         returns (uint256);
 
-    function exchange(int128 i, int128 j, uint256 dx, uint256 min_dy, bool use_eth)
+    function exchange(uint256 i, uint256 j, uint256 dx, uint256 min_dy, bool use_eth)
         external
         payable
         returns (uint256);
